File structure change for easier developement flow
