+++ lib/LED_clock/Config/Animations/clock-only/README.md
@@ -1,0 +1,1 @@
+# Animations designed for a fully featured 24h display with intermediate segments between the digits
