+++ lib/LED_clock/Config/Setup/clock-only/Configuration.h
@@ -56,7 +56,7 @@
 		/**
 		 * @brief The host name that shall be used for OTA updates. If you change this here it must also be changed in the platformio.ini file
 		 */
-		#define OTA_UPDATE_HOST_NAME	"LED-Shelf"
+		#define OTA_UPDATE_HOST_NAME	"LED-Clock"
 	#endif
 
 	/**
@@ -93,6 +93,11 @@
  * @brief Color of the internal LEDs, this will be the default color if blynk functionality is disabled
  */
 #define INTERNAL_COLOR						CRGB::Blue
+
+/**
+ * @brief Color of the seperation dot LEDs, this will be the default color if blynk functionality is disabled
+ */
+#define SEPERATION_DOT_COLOR				CRGB::Blue
 
 /**
  * @brief Color of the LEDs for the OTA update progress bar
@@ -164,22 +169,22 @@
 /**
  * @brief Total number of segments that have LEDs in the shelf
  */
-#define NUM_SEGMENTS 			23
+#define NUM_SEGMENTS 			37
 
 /**
  * @brief Number of LEDs in each segment
  */
-#define NUM_LEDS_PER_SEGMENT	9
+#define NUM_LEDS_PER_SEGMENT	8
 
 /**
  * @brief If you wired the down lighter LEDs to the end of the LED strips set this to true
  */
-#define APPEND_DOWN_LIGHTERS	false
+#define APPEND_DOWN_LIGHTERS	true
 
 /**
  * @brief Number of LEDs For interrior lights
  */
-#define ADDITIONAL_LEDS			12
+#define ADDITIONAL_LEDS			0
 
 /**
  * @brief Automatically calculated total number of LEDs used
@@ -198,13 +203,21 @@
 /**
  * @brief Number of displays in the shelf
  */
-#define NUM_DISPLAYS			4
+#define NUM_DISPLAYS			7
 
 /**
  * @brief These enum definitions are used in the code do address the different Seven segment displays.
  * 		  The numbers have to match with the place of the display in the SegmentDisplayModes array in the file DisplayConfiguration.cpp
  */
-enum DisplayIDs {LOWER_DIGIT_MINUTE_DISPLAY = 0, HIGHER_DIGIT_MINUTE_DISPLAY = 1, LOWER_DIGIT_HOUR_DISPLAY = 2, HIGHER_DIGIT_HOUR_DISPLAY = 3};
+enum DisplayIDs {
+	HIGHER_DIGIT_HOUR_DISPLAY = 0,
+	FIRST_INTERMEDIATE_DISPLAY = 1,
+	LOWER_DIGIT_HOUR_DISPLAY = 2,
+	SECOND_INTERMEDIATE_DISPLAY = 3,
+	HIGHER_DIGIT_MINUTE_DISPLAY = 4,
+	THIRD_INTERMEDIATE_DISPLAY = 5,
+	LOWER_DIGIT_MINUTE_DISPLAY = 6
+};
 
 /**
  * @brief If set to true the display will show 0 at midnight and 12 otherwise
@@ -219,7 +232,7 @@ enum DisplayIDs {LOWER_DIGIT_MINUTE_DISPLAY = 0, HIGHER_DIGIT_MINUTE_DISPLAY = 1
 /**
  * @brief If set to true 24 hour format will be used. For this one additional column is needed in the shelf to display it correctly
  */
-#define USE_24_HOUR_FORMAT			false
+#define USE_24_HOUR_FORMAT			true
 
 /**
  * @brief The number of segments to use for displaying a progress bar for the OTA updates
@@ -239,7 +252,7 @@ enum DisplayIDs {LOWER_DIGIT_MINUTE_DISPLAY = 0, HIGHER_DIGIT_MINUTE_DISPLAY = 1
 /**
  * @brief If set to -1 the flashing middle dot is disabled, otherwise this is the index of the Display segment that should display the dot.
  */
-#define DISPLAY_FOR_SEPERATION_DOT -1
+#define DISPLAY_FOR_SEPERATION_DOT SECOND_INTERMEDIATE_DISPLAY
 
 /**
  * @brief Target Frames per second for the smoothness of animations
@@ -252,6 +265,21 @@ enum DisplayIDs {LOWER_DIGIT_MINUTE_DISPLAY = 0, HIGHER_DIGIT_MINUTE_DISPLAY = 1
  */
 #define ANIMATION_AFTERGLOW			0.2
 
+/**
+ * @brief Lenght of the dot/s fading animation. One flash fades in and out
+ */
+#define DOT_FLASH_SPEED 2000
+
+/**
+ * @brief Intervall in which the dot/s should flash
+ */
+#define DOT_FLASH_INTERVALL	4000
+
+/**
+ * @brief Number of seperation dots to use by default (or if no blynk functionality is avaliable)
+ * 		  allowed values are 1, 2 and 0 to turn it off
+ */
+#define NUM_SEPERATION_DOTS	2
 
 /***************************
 *
@@ -262,7 +290,7 @@ enum DisplayIDs {LOWER_DIGIT_MINUTE_DISPLAY = 0, HIGHER_DIGIT_MINUTE_DISPLAY = 1
 /**
  * @brief Enable automatic brightness adjustments based on a light sensor
  */
-#define ENABLE_LIGHT_SENSOR			true
+#define ENABLE_LIGHT_SENSOR			false
 
 #if ENABLE_LIGHT_SENSOR == true
 	/**
