+++ lib/LED_clock/Config/Setup/clock-only/DisplayConfiguration.cpp
@@ -126,11 +126,9 @@ uint8_t DisplayManager::diplayIndex[NUM_SEGMENTS] = {
 	HIGHER_DIGIT_HOUR_DISPLAY,
 	HIGHER_DIGIT_HOUR_DISPLAY,
 	HIGHER_DIGIT_HOUR_DISPLAY,
-
 	FIRST_INTERMEDIATE_DISPLAY,
 	FIRST_INTERMEDIATE_DISPLAY,
 	FIRST_INTERMEDIATE_DISPLAY,
-
 	LOWER_DIGIT_HOUR_DISPLAY,
 	LOWER_DIGIT_HOUR_DISPLAY,
 	LOWER_DIGIT_HOUR_DISPLAY,
@@ -138,11 +136,9 @@ uint8_t DisplayManager::diplayIndex[NUM_SEGMENTS] = {
 	LOWER_DIGIT_HOUR_DISPLAY,
 	LOWER_DIGIT_HOUR_DISPLAY,
 	LOWER_DIGIT_HOUR_DISPLAY,
-
 	SECOND_INTERMEDIATE_DISPLAY,
 	SECOND_INTERMEDIATE_DISPLAY,
 	SECOND_INTERMEDIATE_DISPLAY,
-
 	HIGHER_DIGIT_MINUTE_DISPLAY,
 	HIGHER_DIGIT_MINUTE_DISPLAY,
 	HIGHER_DIGIT_MINUTE_DISPLAY,
@@ -150,11 +146,9 @@ uint8_t DisplayManager::diplayIndex[NUM_SEGMENTS] = {
 	HIGHER_DIGIT_MINUTE_DISPLAY,
 	HIGHER_DIGIT_MINUTE_DISPLAY,
 	HIGHER_DIGIT_MINUTE_DISPLAY,
-
 	THIRD_INTERMEDIATE_DISPLAY,
 	THIRD_INTERMEDIATE_DISPLAY,
 	THIRD_INTERMEDIATE_DISPLAY,
-
 	LOWER_DIGIT_MINUTE_DISPLAY,
 	LOWER_DIGIT_MINUTE_DISPLAY,
 	LOWER_DIGIT_MINUTE_DISPLAY,
