+++ lib/LED_clock/Config/Setup/diy-machines/DisplayConfiguration.cpp
@@ -1,104 +1,64 @@
 #include "DisplayManager.h"
-// Configuration for the whole LED setup.
+// Configuration the whole LED setup.
 
 /**
  * @brief Each segment belongs to one some Display. This array defines the segment position within this one display.
  * 		  The order of these has to mach the order in which the LEDs are wired
  */
 SevenSegment::SegmentPosition DisplayManager::SegmentPositions[NUM_SEGMENTS] = {
-	SevenSegment::MiddleBottomSegment,
-	SevenSegment::RightBottomSegment,
 	SevenSegment::RightTopSegment,
 	SevenSegment::MiddleTopSegment,
 	SevenSegment::LeftTopSegment,
-	SevenSegment::LeftBottomSegment,
 	SevenSegment::CenterSegment,
-
-	SevenSegment::CenterSegment,
-	SevenSegment::MiddleTopSegment,
-	SevenSegment::MiddleBottomSegment,
-
-	SevenSegment::MiddleBottomSegment,
 	SevenSegment::RightBottomSegment,
+	SevenSegment::MiddleBottomSegment,
+	SevenSegment::LeftBottomSegment,
 	SevenSegment::RightTopSegment,
 	SevenSegment::MiddleTopSegment,
 	SevenSegment::LeftTopSegment,
-	SevenSegment::LeftBottomSegment,
 	SevenSegment::CenterSegment,
-
-	SevenSegment::CenterSegment,
-	SevenSegment::MiddleTopSegment,
-	SevenSegment::MiddleBottomSegment,
-
-	SevenSegment::MiddleBottomSegment,
 	SevenSegment::RightBottomSegment,
+	SevenSegment::MiddleBottomSegment,
+	SevenSegment::LeftBottomSegment,
 	SevenSegment::RightTopSegment,
 	SevenSegment::MiddleTopSegment,
 	SevenSegment::LeftTopSegment,
-	SevenSegment::LeftBottomSegment,
 	SevenSegment::CenterSegment,
-
-	SevenSegment::CenterSegment,
-	SevenSegment::MiddleTopSegment,
-	SevenSegment::MiddleBottomSegment,
-
-	SevenSegment::MiddleBottomSegment,
 	SevenSegment::RightBottomSegment,
-	SevenSegment::RightTopSegment,
-	SevenSegment::MiddleTopSegment,
-	SevenSegment::LeftTopSegment,
+	SevenSegment::MiddleBottomSegment,
 	SevenSegment::LeftBottomSegment,
-	SevenSegment::CenterSegment
+	SevenSegment::RightTopSegment,
+	SevenSegment::RightBottomSegment
 };
 
 /**
- * @brief Each segemnt has a direction, this is important for animation.
+ * @brief Each segemnt has a direction, this is important for animation. 
  * 		  The order of them is important and the direction has to mach the sequence in which the LEDs are wired
  */
 Segment::direction DisplayManager::SegmentDirections[NUM_SEGMENTS] = {
+	Segment::BOTTOM_TO_TOP,
+	Segment::RIGHT_TO_LEFT,
+	Segment::TOP_TO_BOTTTOM,
 	Segment::LEFT_TO_RIGHT,
+	Segment::TOP_TO_BOTTTOM,
+	Segment::RIGHT_TO_LEFT,
 	Segment::BOTTOM_TO_TOP,
 	Segment::BOTTOM_TO_TOP,
 	Segment::RIGHT_TO_LEFT,
 	Segment::TOP_TO_BOTTTOM,
+	Segment::LEFT_TO_RIGHT,
 	Segment::TOP_TO_BOTTTOM,
-	Segment::LEFT_TO_RIGHT,
-
-	Segment::LEFT_TO_RIGHT,
 	Segment::RIGHT_TO_LEFT,
-	Segment::LEFT_TO_RIGHT,
-
-	Segment::LEFT_TO_RIGHT,
 	Segment::BOTTOM_TO_TOP,
 	Segment::BOTTOM_TO_TOP,
 	Segment::RIGHT_TO_LEFT,
 	Segment::TOP_TO_BOTTTOM,
+	Segment::LEFT_TO_RIGHT,
 	Segment::TOP_TO_BOTTTOM,
-	Segment::LEFT_TO_RIGHT,
-
-	Segment::LEFT_TO_RIGHT,
 	Segment::RIGHT_TO_LEFT,
-	Segment::LEFT_TO_RIGHT,
-
-	Segment::LEFT_TO_RIGHT,
 	Segment::BOTTOM_TO_TOP,
-	Segment::BOTTOM_TO_TOP,
-	Segment::RIGHT_TO_LEFT,
 	Segment::TOP_TO_BOTTTOM,
-	Segment::TOP_TO_BOTTTOM,
-	Segment::LEFT_TO_RIGHT,
-
-	Segment::LEFT_TO_RIGHT,
-	Segment::RIGHT_TO_LEFT,
-	Segment::LEFT_TO_RIGHT,
-
-	Segment::LEFT_TO_RIGHT,
-	Segment::BOTTOM_TO_TOP,
-	Segment::BOTTOM_TO_TOP,
-	Segment::RIGHT_TO_LEFT,
-	Segment::TOP_TO_BOTTTOM,
-	Segment::TOP_TO_BOTTTOM,
-	Segment::LEFT_TO_RIGHT
+	Segment::TOP_TO_BOTTTOM
 };
 
 /**
@@ -106,12 +66,9 @@ Segment::direction DisplayManager::SegmentDirections[NUM_SEGMENTS] = {
  */
 SevenSegment::SevenSegmentMode DisplayManager::SegmentDisplayModes[NUM_DISPLAYS] = {
 	SevenSegment::FULL_SEGMENT,
-	SevenSegment::HALF_SEGMENT,
 	SevenSegment::FULL_SEGMENT,
-	SevenSegment::HALF_SEGMENT,
 	SevenSegment::FULL_SEGMENT,
-	SevenSegment::HALF_SEGMENT,
-	SevenSegment::FULL_SEGMENT
+	SevenSegment::ONLY_ONE
 };
 
 /**
@@ -119,41 +76,27 @@ SevenSegment::SevenSegmentMode DisplayManager::SegmentDisplayModes[NUM_DISPLAYS]
  * 		  They define which segment belongs to which Display in the order that they are wired in
  */
 uint8_t DisplayManager::diplayIndex[NUM_SEGMENTS] = {
-	0,
-	0,
-	0,
-	0,
-	0,
-	0,
-	0,
-	1,
-	1,
-	1,
-	2,
-	2,
-	2,
-	2,
-	2,
-	2,
-	2,
-	3,
-	3,
-	3,
-	4,
-	4,
-	4,
-	4,
-	4,
-	4,
-	4,
-	5,
-	5,
-	5,
-	6,
-	6,
-	6,
-	6,
-	6,
-	6,
-	6
+	LOWER_DIGIT_MINUTE_DISPLAY,
+	LOWER_DIGIT_MINUTE_DISPLAY,
+	LOWER_DIGIT_MINUTE_DISPLAY,
+	LOWER_DIGIT_MINUTE_DISPLAY,
+	LOWER_DIGIT_MINUTE_DISPLAY,
+	LOWER_DIGIT_MINUTE_DISPLAY,
+	LOWER_DIGIT_MINUTE_DISPLAY,
+	HIGHER_DIGIT_MINUTE_DISPLAY,
+	HIGHER_DIGIT_MINUTE_DISPLAY,
+	HIGHER_DIGIT_MINUTE_DISPLAY,
+	HIGHER_DIGIT_MINUTE_DISPLAY,
+	HIGHER_DIGIT_MINUTE_DISPLAY,
+	HIGHER_DIGIT_MINUTE_DISPLAY,
+	HIGHER_DIGIT_MINUTE_DISPLAY,
+	LOWER_DIGIT_HOUR_DISPLAY,
+	LOWER_DIGIT_HOUR_DISPLAY,
+	LOWER_DIGIT_HOUR_DISPLAY,
+	LOWER_DIGIT_HOUR_DISPLAY,
+	LOWER_DIGIT_HOUR_DISPLAY,
+	LOWER_DIGIT_HOUR_DISPLAY,
+	LOWER_DIGIT_HOUR_DISPLAY,
+	HIGHER_DIGIT_HOUR_DISPLAY,
+	HIGHER_DIGIT_HOUR_DISPLAY
 };
