--- lib/LED_clock/Config/clock-only.md
@@ -1,3 +1,0 @@
-# Configuration for a fully loaded 24h display with intermediate segments
-Wiring diagram for this version:
-![]()
