+++ lib/LED_clock/README.md
@@ -1,0 +1,4 @@
+# Setup instructions
+To choose which version you want to build simply rename the matching .json file in this directory to library.json
+
+For example: To use the default version simply rename library_default.json to library.json
