--- lib/LED_clock/library.json
@@ -1,28 +1,0 @@
-{
-    "name": "LED_Clock-Lib",
-    "version": "1.0.0",
-    "build": {
-		"srcFilter": [
-			"+<*.c>",
-            "+<*.cpp>",
-            "+<**/*.cpp>",
-            "+<**/*.c>",
-			"+<**/**/*.cpp>",
-            "+<**/**/*.c>",
-            "+<**/**/**/*.cpp>",
-			"+<**/**/**/*.c>"
-		],
-		"flags": [
-            "-I Config",
-            "-I Config/Blynk",
-            "-I Config/Animations",
-            "-I Config/Transitions",
-			"-I Modules/Animator/inc",
-            "-I Modules/ClockState/inc",
-            "-I Modules/DisplayManager/inc",
-            "-I Modules/Misc/inc",
-            "-I Modules/SevenSegment/inc",
-            "-I Modules/TimeManager/inc"
-		]
-	}
-}
