+++ lib/LED_clock/library_diy-machines.json
@@ -1,0 +1,32 @@
+{
+    "name": "LED_Clock-DIY-machines",
+    "version": "1.0.0",
+    "build": {
+		"srcFilter": [
+			"+<Modules/*.c>",
+            "+<Modules/*.cpp>",
+            "+<Modules/**/*.cpp>",
+            "+<Modules/**/*.c>",
+            "+<Config/Animations/diy-machines/*.c>",
+            "+<Config/Animations/diy-machines/*.cpp>",
+            "+<Config/Blynk/default/*.c>",
+            "+<Config/Blynk/default/*.cpp>",
+            "+<Config/Setup/diy-machines/*.c>",
+			"+<Config/Setup/diy-machines/*.cpp>",
+            "+<Config/Transitions/default/*.c>",
+            "+<Config/Transitions/default/*.cpp>"
+		],
+		"flags": [
+            "-I Modules/Animator/inc",
+            "-I Modules/ClockState/inc",
+            "-I Modules/DisplayManager/inc",
+            "-I Modules/Misc/inc",
+            "-I Modules/SevenSegment/inc",
+            "-I Modules/TimeManager/inc",
+            "-I Config/Setup/diy-machines",
+            "-I Config/Blynk/default",
+            "-I Config/Animations/diy-machines",
+            "-I Config/Transitions/default"
+		]
+	}
+}
